add: implementei a passagem de telas agora incluindo a tela de foto
