+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools" >
 
     <application
         android:allowBackup="true"
@@ -10,12 +10,9 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/Theme.DenTeeth"
-        tools:targetApi="31">
+        tools:targetApi="31" >
         <activity
             android:name=".TelaCameraActivity"
-            android:exported="false" />
-        <activity
-            android:name=".FotoPerfilActivity"
             android:exported="false" />
         <activity
             android:name=".RegistroActivity"
@@ -25,7 +22,7 @@
             android:exported="false" />
         <activity
             android:name=".TelaInicioActivity"
-            android:exported="true">
+            android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
