+++ app/src/main/res/layout/activity_tela_camera.xml
@@ -31,7 +31,7 @@
         />
 
     <androidx.camera.view.PreviewView
-        android:id="@+id/CameraPreview"
+        android:id="@+id/pvViewFinder"
         android:layout_width="match_parent"
         android:layout_height="360dp"
         app:layout_constraintTop_toBottomOf="@id/barra"
@@ -47,11 +47,12 @@
         android:textSize="15sp"
         android:fontFamily="@font/inter_regular"
         android:textColor="@color/black"
-        app:layout_constraintTop_toBottomOf="@+id/CameraPreview"
+        app:layout_constraintTop_toBottomOf="@+id/pvViewFinder"
+        app:layout_constraintStart_toStartOf="parent"
         />
 
     <com.google.android.material.button.MaterialButton
-        android:id="@+id/mbtnCapture"
+        android:id="@+id/btnCapture"
         android:layout_width="78dp"
         android:layout_height="80dp"
         android:elevation="2dp"
@@ -70,12 +71,12 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:text="Deixar para depois"
-        android:textSize="19dp"
+        android:textSize="19sp"
         android:layout_marginTop="17dp"
         android:backgroundTint="@color/black"
         android:textColor="@color/white"
         android:fontFamily="@font/inter_regular"
-        app:layout_constraintTop_toBottomOf="@+id/mbtnCapture"
+        app:layout_constraintTop_toBottomOf="@+id/btnCapture"
         />
